requirements.txt
fastapi
uvicorn
pydantic
pandas
xlrd
python-dotenv
google.generativeai
fuzzywuzzy
spacy
numpy
